+++ integration-testing/Cargo.toml
@@ -11,6 +11,8 @@ env_logger = "0.10.0"
 anyhow = { workspace = true }
 ipc-sdk = { workspace = true }
 log = { workspace = true }
+fvm_shared = { workspace = true }
 serde_json = { workspace = true }
+serde = { workspace = true }
 ipc-agent = { path = ".." }
 ipc-identity = { path = "../identity" }
